% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_portal}
\alias{browse_portal}
\title{Open the City of Toronto Open Data Portal in your browser}
\usage{
browse_portal()
}
\value{
A browser is opened to the City of Toronto Open Data Portal website
if the session is interactive. The URL is returned as a character string.
}
\description{
Opens a browser to \code{https://open.toronto.ca}.
}
\examples{
\dontrun{
browse_portal()
}
}
