% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resource.R
\name{download_resource}
\alias{download_resource}
\title{Download a resource}
\usage{
download_resource(url, format, store = "session", path = "file")
}
\arguments{
\item{url}{URL for the resource}

\item{format}{Format of the resource}

\item{store}{One of session (default) or disk. session stores in R session, and disk saves the file to disk.}

\item{path}{If store = disk, you must give a path to store the file to}
}
\description{
Download a resource
}
