% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_resources.R
\name{get_package_resources}
\alias{get_package_resources}
\title{Get Resources for a Package}
\usage{
get_package_resources(package_id)
}
\arguments{
\item{package_id}{Package ID}
}
\description{
Get Resources for a Package
}
