% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages.R
\name{get_packages}
\alias{get_packages}
\title{List packages.}
\usage{
get_packages(limit = 31)
}
\arguments{
\item{limit}{The maximum number of packages to return (default is 31).}
}
\description{
List packages.
}
\examples{
\dontrun{
get_packages(5)
}
}
