% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Download a resource into your R session}
\usage{
get_resource(url, format, store = "session", path = "file")
}
\arguments{
\item{url}{URL of the resource}

\item{format}{Format of the resource}

\item{store}{One of session (default) or disk. session stores in R session, and disk saves the file to disk.}

\item{path}{If store = disk, you must give a path to store the file to}
}
\description{
Download a resource into your R session
}
\examples{
\dontrun{
res <- list_package_resources("832a5967-15a8-4dce-aad5-36d2d45f9d04")
list_resource(url = res[1, "url"], format = res[1, "format"])
}
}
