% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_package_resources.R
\name{list_package_resources}
\alias{list_package_resources}
\title{List resources for a package}
\usage{
list_package_resources(package_id)
}
\arguments{
\item{package_id}{The ID of the package whose resources you are listing.}
}
\description{
List resources for a package
}
\examples{
\dontrun{
list_package_resources("1db34737-ffad-489d-a590-9171d500d453")
}
}
