% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_portal.R
\name{open_portal}
\alias{open_portal}
\title{Open the City of Toronto Open Data Portal in your browser}
\usage{
open_portal(browser = getOption("browser"))
}
\description{
Opens a browser to \code{https://open.toronto.ca}.
}
\examples{
\dontrun{
open_portal()
}
}
