% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendatatoronto.R
\docType{package}
\name{opendatatoronto-package}
\alias{opendatatoronto}
\alias{opendatatoronto-package}
\title{opendatatoronto}
\description{
opendatatoronto is an R interface to the \href{https://open.toronto.ca/}{City of Toronto's Open Data Portal}.
The goal of the package is to help read data directly into R without needing
to manually download it via the portal.
}
\details{
For more information, please start with the vignettes: \code{browseVignettes(package = "opendatatoronto")} or visit the \href{https://sharlagelfand.github.io/opendatatoronto/}{package website}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sharlagelfand.github.io/opendatatoronto/}
  \item \url{https://github.com/sharlagelfand/opendatatoronto/}
  \item \url{https://sharlagelfand.github.io/opendatatoronto}
  \item Report bugs at \url{https://github.com/sharlagelfand/opendatatoronto/issues}
}

}
\author{
\strong{Maintainer}: Sharla Gelfand \email{sharla.gelfand@gmail.com}

Other contributors:
\itemize{
  \item City of Toronto [copyright holder, funder]
}

}
\keyword{internal}
