% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_id_from_url.R
\name{package_id_from_url}
\alias{package_id_from_url}
\title{Get package ID from package portal URL}
\usage{
package_id_from_url(package_url)
}
\arguments{
\item{package_url}{Package URL from the City of Toronto Open Data Portal}
}
\description{
Get the \code{package_id} associated with a given package's URL on the City of Toronto Open Data Portal.
}
\examples{
\dontrun{
package_id_from_url("https://open.toronto.ca/dataset/ttc-subway-delay-data")
}
}
