% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_packages.R
\name{search_packages}
\alias{search_packages}
\title{Search packages by title}
\usage{
search_packages(title, limit = 50)
}
\arguments{
\item{title}{Title to search (case-insensitive).}

\item{limit}{Maximum number of packages to return. The default is 50. The maximum limit is 1000.}
}
\description{
Search portal packages by title.
}
\examples{
\dontrun{
search_packages("ttc")
}
}
