% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{show_package}
\alias{show_package}
\title{Show a package's metadata}
\usage{
show_package(package)
}
\arguments{
\item{package}{A way to identify the package. Either a package ID (passed as a character vector directly) or the package's URL from the portal.}
}
\value{
A tibble including \code{title}, \code{id}, \code{topics},
\code{civic_issues}, \code{dataset_category}, \code{num_resources}
(the number of resources in the package), \code{formats} (the different
formats of the resources), \code{refresh_rate} (how often the package is
refreshed), and \code{last_refreshed} (the date it was last refreshed).
}
\description{
Show a portal package's metadata.
}
\examples{
\donttest{
show_package("c01c6d71-de1f-493d-91ba-364ce64884ac")
}
}
