% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_portal.R
\name{browse_portal}
\alias{browse_portal}
\title{Open the City of Toronto Open Data Portal in your browser}
\usage{
browse_portal()
}
\description{
Opens a browser to \code{https://open.toronto.ca}.
}
\examples{
\dontrun{
browse_portal()
}
}
