% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages.R
\name{get_packages}
\alias{get_packages}
\title{List packages.}
\usage{
get_packages(limit = 31)
}
\arguments{
\item{limit}{(numeric) The maximum number of packages to return (optional, default: 31)}
}
\description{
List packages.
}
