% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_package_resources.R
\name{list_package_resources}
\alias{list_package_resources}
\title{List resources for a package}
\usage{
list_package_resources(package)
}
\arguments{
\item{package}{A way to identify the package. Either a package ID (passed as a character vector directly), a single package resulting from \code{\link{list_packages}} or \code{\link{search_packages}}, or the package's URL from the portal.}
}
\value{
A tibble of resources along with metadata, including \code{name},
\code{id}, \code{format} (the format of the resource file), and
\code{last_modified} (the date the resource was last modified).
}
\description{
List resources for a package on the portal.
}
\examples{
\dontrun{
list_package_resources("1db34737-ffad-489d-a590-9171d500d453")
list_package_resources("https://open.toronto.ca/dataset/ttc-subway-delay-data")
}
}
