% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_package_resources.R
\name{list_package_resources}
\alias{list_package_resources}
\title{List resources for a package}
\usage{
list_package_resources(package)
}
\arguments{
\item{package}{A way to identify the package. Either a package ID (passed as a character vector directly), a single package resulting from \code{\link{list_packages}} or \code{\link{search_packages}}, or the package's URL from the portal}
}
\description{
List resources for a package
}
\examples{
\dontrun{
list_package_resources("1db34737-ffad-489d-a590-9171d500d453")

list_package_resources("https://open.toronto.ca/dataset/ttc-subway-delay-data")
}
}
