% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List packages}
\usage{
list_packages(limit = 50)
}
\arguments{
\item{limit}{The maximum number of packages to return. The default is 50.}
}
\description{
List packages
}
\examples{
\dontrun{
list_packages(5)
}
}
