% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_packages.R
\name{search_packages}
\alias{search_packages}
\title{Search packages by title}
\usage{
search_packages(title, limit = 1000)
}
\arguments{
\item{title}{Title to search (case-insensitive).}

\item{limit}{Maximum number of packages to return. The default (and the maximum possible value) is 1000.}
}
\description{
Search packages by title. Returns a tibble of matching packages along with package metadata, including \code{title}, \code{id}, \code{topics}, \code{dataset_category}, \code{num_resources} (the number of resources in the package), \code{formats} (the different formats of the resources), \code{refresh_rate} (how often the package is refreshed), and \code{last_refreshed} (the date it was last refreshed).
}
\examples{
\dontrun{
search_packages("ttc")
}
}
