% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_packages.R
\name{search_packages}
\alias{search_packages}
\title{Search packages by title}
\usage{
search_packages(title, limit = 50)
}
\arguments{
\item{title}{Title to search (case-insensitive).}

\item{limit}{Maximum number of packages to return. The default is 50. The maximum limit is 1000.}
}
\value{
A tibble of matching packages along with package metadata, including
\code{title}, \code{id}, \code{topics}, \code{civic_issues},
\code{dataset_category}, \code{num_resources} (the number of resources in the
package), \code{formats} (the different formats of the resources),
\code{refresh_rate} (how often the package is refreshed), and
\code{last_refreshed} (the date it was last refreshed).
}
\description{
Search portal packages by title.
}
\examples{
\donttest{
search_packages("ttc")
}
}
