% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_package.R
\name{show_package}
\alias{show_package}
\title{Show information on a package}
\usage{
show_package(package)
}
\arguments{
\item{package}{A way to identify the package. Either a package ID (passed as a character vector directly) or the package's URL from the portal}
}
\description{
Show information on a package
}
\examples{
\dontrun{
show_package("c01c6d71-de1f-493d-91ba-364ce64884ac")
}
}
